requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
dj-database-url==2.1.0
Django==3.2.1
django-countries==7.5.1
django-multiforloop==0.3.1
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
Pillow==9.5.0
protobuf==3.20.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2018.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.27.1
whitenoise==6.4.0
zope.interface==6.1
